requirements.txt
aiogram==2.20
aiohttp==3.8.1
aioschedule==0.5.2
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
psycopg2-binary==2.9.3
python-decouple==3.6
pytz==2022.1
requests==2.28.0
soupsieve==2.3.2.post1
urllib3==1.26.9
yarl==1.7.2
